(REF) Move test for `$translationMode` to its own class
The functionality is bigger than `MessageTemplate`, and the test is fairly long.
