+++ tests/phpunit/CRM/Core/BAO/MessageTemplateTest.php
@@ -3,7 +3,6 @@
 use Civi\Api4\Address;
 use Civi\Api4\Contact;
 use Civi\Api4\MessageTemplate;
-use Civi\Api4\Translation;
 use Civi\Token\TokenProcessor;
 
 /**
@@ -200,56 +199,6 @@ class CRM_Core_BAO_MessageTemplateTest extends CiviUnitTestCase {
     $this->assertEquals('[case #' . $msg->getIdHash() . '] Test 123', $subject);
     $this->assertStringContainsString('Your Case Role(s) : Sand grain counter', $message);
     $this->assertStringContainsString('Case ID : 1234', $message);
-  }
-
-  public function getTranslationSettings(): array {
-    $es = [];
-    $es['fr_FR-full'] = [
-      ['partial_locales' => FALSE, 'uiLanguages' => ['en_US', 'fr_FR', 'fr_CA']],
-    ];
-    $es['fr_FR-partial'] = [
-      ['partial_locales' => TRUE, 'uiLanguages' => ['en_US']],
-    ];
-    return $es;
-  }
-
-  /**
-   * Test that translated strings are rendered for templates where they exist.
-   *
-   * @dataProvider getTranslationSettings
-   * @throws \API_Exception|\CRM_Core_Exception
-   */
-  public function testGetTranslatedTemplate($translationSettings): void {
-    $cleanup = \CRM_Utils_AutoClean::swapSettings($translationSettings);
-
-    $this->individualCreate(['preferred_language' => 'fr_FR']);
-    $this->contributionCreate(['contact_id' => $this->ids['Contact']['individual_0']]);
-    $this->addTranslation();
-
-    $messageTemplate = MessageTemplate::get()
-      ->addWhere('is_default', '=', 1)
-      ->addWhere('workflow_name', 'IN', ['contribution_online_receipt', 'contribution_offline_receipt'])
-      ->addSelect('id', 'msg_subject', 'msg_html', 'workflow_name')
-      ->setLanguage('fr_FR')
-      ->setTranslationMode('fuzzy')
-      ->execute()->indexBy('workflow_name');
-
-    $this->assertFrenchTranslationRetrieved($messageTemplate['contribution_online_receipt']);
-
-    $this->assertStringContainsString('{ts}Contribution Receipt{/ts}', $messageTemplate['contribution_offline_receipt']['msg_subject']);
-    $this->assertStringContainsString('Below you will find a receipt', $messageTemplate['contribution_offline_receipt']['msg_html']);
-    $this->assertArrayNotHasKey('actual_language', $messageTemplate['contribution_offline_receipt']);
-
-    $messageTemplate = MessageTemplate::get()
-      ->addWhere('is_default', '=', 1)
-      ->addWhere('workflow_name', 'IN', ['contribution_online_receipt', 'contribution_offline_receipt'])
-      ->addSelect('id', 'msg_subject', 'msg_html', 'workflow_name')
-      ->setLanguage('fr_CA')
-      ->setTranslationMode('fuzzy')
-      ->execute()->indexBy('workflow_name');
-
-    $this->assertFrenchTranslationRetrieved($messageTemplate['contribution_online_receipt']);
-
   }
 
   /**
@@ -969,40 +918,6 @@ id |' . $tokenData['contact_id'] . '
 t_stuff.favourite_emoticon |
 ';
     return $expected;
-  }
-
-  /**
-   * @return mixed
-   * @throws \API_Exception
-   * @throws \Civi\API\Exception\UnauthorizedException
-   */
-  private function addTranslation() {
-    $messageTemplateID = MessageTemplate::get()
-      ->addWhere('is_default', '=', 1)
-      ->addWhere('workflow_name', '=', 'contribution_online_receipt')
-      ->addSelect('id')
-      ->execute()->first()['id'];
-
-    Translation::save()->setRecords([
-      ['entity_field' => 'msg_subject', 'string' => 'Bonjour'],
-      ['entity_field' => 'msg_html', 'string' => 'Voila!'],
-      ['entity_field' => 'msg_text', 'string' => '{contribution.total_amount}'],
-    ])->setDefaults([
-      'entity_table' => 'civicrm_msg_template',
-      'entity_id' => $messageTemplateID,
-      'status_id:name' => 'active',
-      'language' => 'fr_FR',
-    ])->execute();
-    return $messageTemplateID;
-  }
-
-  /**
-   * @param $contribution_online_receipt
-   */
-  private function assertFrenchTranslationRetrieved($contribution_online_receipt): void {
-    $this->assertEquals('Bonjour', $contribution_online_receipt['msg_subject']);
-    $this->assertEquals('Voila!', $contribution_online_receipt['msg_html']);
-    $this->assertEquals('fr_FR', $contribution_online_receipt['actual_language']);
   }
 
 }
